refactor(auth): merge the browser satellites into browser_capture (ADR-0033 PR 4.1)
`browser_state_validation.py` (56) and `login_wait_trace.py` (181) existed only
because of the ADR-0008 module-size cap, and both said so in their own comments:
browser_capture's import site cited ADR-0008 as the reason the validation bridge
was a leaf, and login_wait_trace's docstring said outright it was split out "so
the capture core stays under the ADR-0008 module-size budget".

Neither earns its keep as a leaf. `heal_captured_state` fails the deletion test
outright — both of its callers are inside browser_capture and its body just
adapts `httpOnly` -> `http_only` and delegates to the shared PSIDTS recovery.
`login_wait_trace` has exactly one consumer (browser_capture, 3 call sites).
ADR-0033 lifts the cap for `_auth` precisely so a seam like this can be one deep
module instead of three shallow ones.

`browser_launch_errors.py` deliberately does NOT move: `classify_launch_failure`
has a second, independent consumer in `cli/services/login/master_token.py`, so it
passes the deletion test on its own and stays a real leaf.

This commit is the wholesale move only — no behavior change, no simplification.
The bodies are carried verbatim into two labelled sections; the guardrail,
test-seam and docs bookkeeping follows in the next commit, so the ratchet and the
tracing tests are red at this commit alone.

Preserved deliberately:

* `trace_url` stays a SECOND URL redactor alongside `extraction._safe_url`, and
  its docstring now says "do not unify them" in as many words. Its stricter
  policy is the point: it keeps ONLY scheme + host because it renders arbitrary
  SSO redirects, where a federated IdP can carry a one-time assertion in the
  path, whereas `_safe_url` preserves the path outside a Google-OAuth allowlist.
* `log_observed_navigations` stays inert unless DEBUG is on (no listener is ever
  attached) and still swallows every listener exception, so it cannot destabilise
  the five-minute login wait.
* `heal_captured_state` still returns `(state, error)` and never raises, so a
  failed heal cannot discard a completed sign-in.

The absorbed modules become re-export shims with removal notes (next major).
`login_wait_trace`'s note also records the one unavoidable consequence of a
module merge: the tracing helpers now log under the
`notebooklm._auth.browser_capture` logger, since a logger follows its defining
module.

`browser_capture.__all__` drops from 24 to 22. It was hand-simulating a package
interface — several entries existed only to give a CLI caller one sanctioned
import site for a name owned by another leaf. `log_observed_navigations` and
`safe_page_url` are no longer re-exports of anything and have no consumer outside
this module, so they are no longer advertised; every name the CLI actually
imports (`cli/services/playwright_login.py`'s 16, plus `classify_launch_failure`
and `sync_playwright_context` for the master-token bootstrap and
`app_host_scope_note` for the cookie-jar advice) is untouched.

The two collaborator imports (`cookies`, `psidts_recovery`) move from the
absorbed leaf to module scope here. That is graph-neutral: browser_capture
already reached both transitively through the leaf it imported at module scope,
and neither module imports back, so no cycle is created.
